requirements/requirements_dev.txt
ansible  # Used in build process to generate some configs
build
django-extensions
django-debug-toolbar
ipython
tox
typeguard
pytest
pytest-asyncio
pytest-xdist
pytest-cov
pytest-django
setuptools-scm
psycopg[binary]
sdb
sphinx
sphinx-rtd-theme
sphinx-autodoc-typehints
